add annunciator field for train numbers
now there are 3 train numbers that can be forwarded over the block cable
Offer, annunciator and track field numbers.
to transport the LN display message over the block cable the broadcast type of message is used
the payload of the broadcast is the LN message for setting a train number field.
so byte 2 of the broadcast is 0xE5 which in turn is the opc for the LN display message
broadcast: http://fremo-block.sourceforge.net/BlockInterface/MessageBroadcast.de.html
train number display: http://fremo-stw.sourceforge.net/ZNF800/TrainNumberDisplayMessage.en.html