clear "Erlaubnisabgabe" message marker in State "erhalten" when set.
if the box is in state "erhalten" more "Erlaubnisabgabe" messages can
by received. These messages were remembered by a marker.
The problem:
If the box itself sends "Erlaubnisabgabe" to the other box, it sees
the before remembered message an changes the state back to "erhalten".
The result now is the both boxes are in the state "erhalten".
Solution:
Clear the marker for the "Erlaubnisabgabe" message when it is set and
the box is in state "erhalten".